fix(web): code scalability and abstractions
